32bit shift too large for variable warning fixed when compiled with VS 2005 (wildcard_25)
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@4630 3231ed4b-2112-0410-84e0-8c3944a40730

+++ shareaza/DownloadTransferED2K.cpp
@@ -1067,10 +1067,10 @@ BOOL CDownloadTransferED2K::OnSendingPart64(CEDPacket* pPacket)
 	}
 
 	QWORD	nOffset = pPacket->ReadLongLE();
-			nOffset = ( pPacket->ReadLongLE() << 32 ) | nOffset;
+			nOffset = ( (unsigned __int64)pPacket->ReadLongLE() << 32 ) | nOffset;
 	
 	QWORD	nLength = pPacket->ReadLongLE();
-			nLength = ( pPacket->ReadLongLE() << 32 ) | nLength;
+			nLength = ( (unsigned __int64)pPacket->ReadLongLE() << 32 ) | nLength;
 
 	if ( nLength <= nOffset )
 	{
@@ -1126,7 +1126,7 @@ BOOL CDownloadTransferED2K::OnCompressedPart64(CEDPacket* pPacket)
 	}
 
 	QWORD	nBaseOffset = pPacket->ReadLongLE();
-			nBaseOffset = ( pPacket->ReadLongLE() << 32 ) | nBaseOffset;
+			nBaseOffset = ( (unsigned __int64)pPacket->ReadLongLE() << 32 ) | nBaseOffset;
 
 	QWORD	nBaseLength = pPacket->ReadLongLE();	// Length of compressed data is 32bit
